+++ src/modules/settings.js
@@ -1,4 +1,5 @@
 import { moduleId, SETTINGS } from './constants.js';
+import {theme} from "./stores.js"
 
 export let setting = key => {
   return game.settings.get(moduleId, key);
@@ -151,7 +152,7 @@ export function initSettings(app) {
     },
     default: "light",
     type: String,
-    onChange: debouncedReload
+    onChange: v => theme.set(v)
   });
 
   game.settings.register(moduleId, SETTINGS.INVERT_CLICKS, {