+++ src/view/HelpUI.svelte
@@ -5,7 +5,13 @@
    import { ApplicationShell } from "@typhonjs-fvtt/runtime/svelte/component/core";
    import "../main.scss";
    import { helpTopic, helpTree, expanded, selectedHelp, theme } from "../modules/stores.js";
-   import { onDestroy } from "svelte";
+   import { onDestroy, getContext } from "svelte";
+
+   const { application } = getContext("external");
+   const position = application.position;
+   const { height } = position.stores;
+   let contentH = $height;
+   onDestroy(height.subscribe((h) => (contentH = h - 30)));
 
    export let elementRoot;
 
@@ -72,15 +78,17 @@
             />
          </div>
       </div>
-      <div class="ui-bg-base-100 ui-flex-col ui-flex ui-w-[70%] ui-p-2" id="help-page">
-         {#if topic?.component}
-            <svelte:component this={topic.component} />
-         {:else if topic?.content}
-            {@html topic.content}
-         {:else}
-            <h1>Welcome to the Help Center</h1>
-            <p>Please select a page from the left sidebar.</p>
-         {/if}
+      <div class="ui-bg-base-100 ui-flex-col ui-flex ui-w-[70%] ui-p-2" id="help-page" style="height: {contentH}px;">
+         <div class="ui-h-full ui-overflow-y-auto">
+            {#if topic?.component}
+               <svelte:component this={topic.component} />
+            {:else if topic?.content}
+               {@html topic.content}
+            {:else}
+               <h1>Welcome to the Help Center</h1>
+               <p>Please select a page from the left sidebar.</p>
+            {/if}
+         </div>
       </div>
    </main>
 </ApplicationShell>