+++ src/view/components/BrowserSettings.svelte
@@ -10,7 +10,7 @@
 </script>
 
 <div class="ui-bg-base ui-p-2 browser-settings ui-flex ui-flex-col ui-gap-2 ui-h-full">
-   <div class="ui-text-center ui-text-lg ui-font-bold">Settings</div>
+   <div class="ui-text-center ui-text-base-content ui-text-lg ui-font-bold">Settings</div>
    <ArgInput
       type="bool"
       label="Advanced mode"