+++ src/view/help/HelpBrowserSettings.svelte
@@ -1,1 +1,7 @@
-TBD
+<h1>Alpha Browser settings</h1>
+
+<ul>
+   <li>"Advanced mode" is for displaying a Filter Bar in the Filter Panel</li>
+   <li>"Page size" is a number of items on the page</li>
+   <li>"Ignored packs" is for excluding certain compendiums from the views by type</li>
+</ul>